Slide interval endpoints to non-removed segments (#16146)
The test farm exposed a bug where intervals were being created on
removed segments. We originally thought this bug was due to the assigned
referenceTypes on the start and end localReferences, but it was actually
due to the start and endpoints not sliding as a result of removing a
range of text. The fixes included in this pr likely close both
[AB#4543](https://dev.azure.com/fluidframework/internal/_workitems/edit/4543)
and
[AB#4544](https://dev.azure.com/fluidframework/internal/_workitems/edit/4544).
The changeProperties and removeRange edits still fail with
"Non-transient references need segment" errors, but those will be dealt
with in another PR. More investigation is needed to determine if the
[other
bugs](https://dev.azure.com/fluidframework/internal/_workitems/edit/4615)
I previously opened in ADO are still relevant.

---------

Co-authored-by: Scarlett Lee <90647596+scarlettjlee@users.noreply.github.com>