add checkpoints TTL (#16298)
Sets a TTL on the local checkpoints collection by adding a TTL index.
This value is configurable with a default value of 10 days. Checkpoint
data is written to the global database when the session becomes inactive
(no clients), so the data in the local database is no longer the most
recent.