Stress: Clear stashed ops on error (#16347)
Looking at the data it seems we are generally seeing 0x184 in the load
following another failure, like disposed. My hunch is that the error is
leading to a bad stashed ops object, which then leads to the assert.
Regardless, i think it makes sense to clear the stashed ops object on
failure. We likely need to do more work to ensure that the stashed ops
object is only created when the container is healthy.

related to
[AB#4926](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/4926)

---------

Co-authored-by: Tony Murphy <anthonm@microsoft.com>