Fix bugs in merge command (#16338)
The sync-branch workflow in the pipeline at
https://github.com/microsoft/FluidFramework/actions/workflows/sync-branch.yml
is encountering failures. I was able to replicate the error locally, and
here are the errors along with their corresponding solutions.

1. Error: "branch name required" - To fix this, update the reference to
refs/remotes/origin/next instead of just "next" when passing it as a
branch name.

2. Error: "Committer identity unknown" - To resolve this, configure Git
to check the merge operation by setting the global user email and name
using the commands:
`git config --global user.email "you@example.com"`
`git config --global user.name "Your Name"`

3. Error: "could not read Username for 'https://github.com/': No such
device or address" - This error occurred while pushing the branch to the
remote repository. To fix it, set the Git configuration using the
following command, which substitutes the default GitHub URL with a
custom URL that includes a secret token:
`await this.gitRepo.gitClient.raw([
"config",
"--global",
`url.https://${secret}@github.com/.insteadOf`,
"https://github.com/",]);`

In summary, the `url.https://${secret}@github.com/.insteadOf`
configuration is used to substitute the default GitHub URL with a custom
URL that includes a secret token for secure authentication and
authorization during Git operations, such as pushing to a remote
repository on GitHub.

4. The GitHub workflow fails when there is an existing open PR. Fixed
the workflow to not error out.


[AB#4873](https://dev.azure.com/fluidframework/internal/_sprints/taskboard/Team%20Sasha/internal/Team%20Sasha%20-%202023.06-B?workitem=4873)