fix(merge:branches): Switch branches before trying to delete the branch (#16398)
The merge branches command was failing because it was trying to delete
the current branch. I reordered the logic so we switch branches before
deleting the other one.

Adds some hidden flags to bypass parts of the command process like the
PR check and the opening of the PR. This makes it easier to test the
command.

Also centralizes the branch cleanup and calls it even in the success
case, since we still want to clean up the branches on success.