fix(merge:branches): Merge source into target branch instead of the other way (#16496)
In the existing logic and recommended steps for developers to follow, we
merge `next` into `main`. While the resulting merge commit is the same
regardless of direction, the direction _does_ matter when using tools to
review history, especially visual tools (e.g. GitKraken, SourceTree,
etc.). See [this StackOverflow
answer](https://stackoverflow.com/a/45973607/551579) for details.

I updated the logic we use to change the direction of the merge. That
will fix the PRs with no conflicts. For conflicting PRs, developers have
to do it themselves I updated the instructions we include with more
details about what to do. They need to check out `next`, then merge that
with the temporary merge branch, then checkout the merge branch again
and fast-forward it to the merge commit. That is:

- `git checkout -b temp-next upstream/next` -- make a temporary branch.
- `git merge upstream/main-next-abc1234` -- merge the main commits into
the next branch.
- Resolve any merge conflicts between the branches, then commit all the
changes
- `git checkout main-next-abc1234` -- check out the mergeBranch locally
- `git merge temp-next --ff-only` -- fast-forward to the merge commit
- `git push upstream` -- and push it to upstream

This is more complicated, but it's necessary to make our commit history
legible.