Capture Full Stacks in Mocha Tests (#16514)
Seeing some sporadic test failures on tinylicious. The errors loose at
least part of the stack making it hard to know exactly what is failing.
I have a hunch we are missing retry logic, having the full stack will
help identify if that is that case. This should also help with debugging
test failures overall.

We already do this for stress and it works well:
https://github.com/microsoft/FluidFramework/blob/be77e843ef5c075c737eea578f91313f27b16f14/packages/test/test-service-load/src/runner.ts#L108

Co-authored-by: Tony Murphy <anthonm@microsoft.com>