Remove nilUuid (#16516)
## Description
NIL is used in experimental/legacy SharedTree for versioning but isn't
needed for the runtime compressor. This matters because consumers of the
container runtime will now end up importing this type and some older
versions of the `uuid` package don't export NIL.


## Reviewer Guidance

The review process is outlined on [this wiki
page](https://github.com/microsoft/FluidFramework/wiki/PR-Guidelines#guidelines).