Prevent interval start > end in revertLocalSequenceRemove (#16509)
In revertLocalSequenceRemove, it's possible to revert interval positions
such that start > end, which is probably not the intent of the undo.
This can happen if only one endpoint was in the deleted text range, and
then another client moved the interval in the meantime. Or if the
interval was already reversed. In this change, it won't move the
interval if that will result in start > end.

I tried to prevent start > end in IntervalCollection.change, but it
broke 126 tests (mostly fuzz) and fixing all of them was taking longer
than a day. So I'm tabling that for now.