Refactor scenario runner to allow shared Client (#15057)
## Description

This work was started and abandoned in #13100.

The goal/requirement here is to allow each stage of ScenarioRunner to
pass along certain parameters to the next stage. This can enable us to
create specific scenarios that rely on multiple things happening with 1
AzureClient instance by using configs rather than building an entire
Runner for each scenario.

For example, a while back we ran into an issue where if 1 AzureClient
instance was used to load 2 different documents, the 2nd document would
fail to load due to cache pollution (see #12060). To repro this
currently with ScenarioRunner, we would have to create an entire new
Runner that spawns 2 documents from 1 AzureClient. Ideally, we could do
like
1. stage: azure-client
    - passes: client 
2. stage: doc-creator
    - consumes: client
    - passes: docId
3. stage: doc-loader
    - consumes: client, docId
4. stage: doc-creator
    - consumes: client
    - passes: docId
5. stage: doc-loader
    - consumes: client, docId

This PR doesn't actually quite get us to this specific repro yet, but
it's 1 step closer.

ScenarioRunner already has `out` env set-up going on, but it currently
always runs in multiple processes. When running in multiple processes,
you can't share an AzureClient instance. This PR just allows tests to be
run with 1 process, rather than multiple processes so that complex
objects can be shared between stages.

## Breaking Changes

- Changes the `IRunner` interface to expect both `run` and `runSync`
(previously just `run`).
- Makes `AzureClientConfig` from each `Runner` shared as
`AzureClientConnectionConfig` from `interfaces.ts`. This is different
from `AzureClientConfig` in `utils.ts`, though many of the properties
mean the same things.

## Reviewer Guidance

Overall, this PR does 3 things:

1. Moves `execRun` from `*RunnerClient.ts` files into each `*Runner.ts`
file as a `public static` class property of the `Runner`. This allows
both `run` and `runSync` to utilize the business logic of the Runner,
and also narrows the `*RunnerClient` scope to parsing command line args
to build a config.
2. Adds a `runSync` method to each `Runner` that operates similarly to
`run`, but using 1 process, rather than spawning child processes.
3. Unifies various types that were duplicated across each Runner for
more consistency and easier boilerplating of new Runners.