Limiting logtail length (#16510)
## Description

We use logtails as an optimization to avoid too many get delta calls.
But, in rare cases, the logtail is growing very big which is causing OOM
restarts in gitrest and breaking some of the scenarios where we use
redis to cache summaries. With this PR, we limit logtails to 2000
(configurable) ops

## Testing

Tested that both client summaries and service summaries without full
logtails will work without issues. The driver ends up making a get delta
call to fetch the trailing ops.