Persist corrupt documents in scribe (#16515)
## Overview
Currently, corrupted document status is stored in memory. When our
microservices restart, we lose this information. In order to prevent
reprocessing of corrupted documents, this change makes it possible if a
document is corrupted in our checkpoints. In this case, we are
persisting if a document is corrupt when there is a protocol error in
scribe.

### Checkpointing
When we mark a document as corrupted in the database, if local
checkpointing is enabled, we will write the updated checkpoint to both
the local and global databases. After the document is marked as corrupt,
we can no longer write checkpoints. Therefore, it is necessary to
persist this information in the global database at this time. Local
checkpoints that are outdated will eventually be removed by the local
collection's TTL.

### Testing
The following scenarios were tested:
- Create a new document that throws a protocol error. Document is marked
as corrupt (in both local and global checkpoints). Other documents are
processed as expected.
- Restart scribe and open a document that is marked as corrupt.
NoOpLambda is created. Other documents are processed as expected.
- Mark an existing document (that was using the old checkpoint schema
without `isCorrupt`) as corrupted