Deprecate ILoaderOptions.cache (#16383)
## Description

The `ILoaderOptions.cache` property has been deprecated. This work is
preparation for completely removing container caching in the Loader.


[AB#3118](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/3118)

## Upcoming Change

### `ILoaderOptions.cache` has been deprecated

`ILoaderOptions.cache` has been deprecated and will be removed in a
future release, as well as all caching functionality of containers.
Cache support will be removed soon, please try not to rely on caching,
and inform us if you cannot do so.