Set loader caching default to false (#16363)
## Description

Loader caching will now be off by default. This work is preparation for
completely removing container caching in the Loader.


[AB#3118](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/3118)

Continuation of: https://github.com/microsoft/FluidFramework/pull/14680

## Breaking Change

### Loader container caching off by default

Loader container caching will now be off by default and the ability to
control it is deprecated. Loader caching is deprecated and will be
removed in a future release, as well as all caching functionality of
containers. Please try not to rely on caching and inform us if you
cannot do so.

If you run into trouble with this behavior, please report it ASAP to the
FluidFramework team and use the following options (available in this
release only) to unblock you:
- set `ILoaderProps.options.cache` to `true` when constructing a
`Loader` object (see the `ILoaderOptions` interface)
- set `[LoaderHeader.cache]` header to `true` when requesting a
container