Fix handling of SharedDirectory ops on subdirs created on a detached data store (#16578)
## Description

`SharedDirectory` was using `-1` all the time for locally created
subdirectories.
This leads to bugs when considering the logic in
`isMessageForCurrentInstanceOfSubDirectory` when the `SharedDirectory`
is additionally detached: no ack op will ever be received for that
subdir (since an op is never sent), but clients who load using the
attachment summary may start sending ops which refers to the same
subdirectory.

merge-tree has a similar problem which is resolved by using `-1` or `0`
depending on the attach state of the DDS, see `UnassignedSequenceNumber`
vs `UniversalSequenceNumber`. The fix here brings SharedDirectory
handling closer to merge-tree's.