Fix maxWaitTime on errors based on whether endpoint(service or socket was reachable or not) (#16328)
## Description

Item: https://dev.azure.com/fluidframework/internal/_workitems/edit/4745
8s of MaxReconnectDelayInMs seems less. Increase it to 30s. This was
done due to a problem we faced recently where "retry-after" was not
present in the Throttling error due to some error at server side and
client was retrying very aggressively as 8s sec wait was too less. So
increase it to 30s in case the service or socket is reachable or not.

The general philosophy behind this PR is the following:
1. If we can't reach service (including being offline), we do not have
any data on when this situation would change, so we want to keep
retrying relatively frequently, to reduce delay from service becoming
reachable to a point client becomes connected and up to date. So,
continue to use back-off strategy with max 8 sec between retries.
2. If we reached service and it returned some retryable error, but there
is no retryAfter property communicated to us, then we will back-off
exponentially with max delay of 30 seconds between retries. This gives
more time for service to recover, and avoids whole application being
throttled due to too many calls performed by all clients.

8 & 30 seconds could be changed in the future, but we want to keep two
different policies for those two distinct cases.

---------

Co-authored-by: Jatin Garg <jatingarg@Jatins-MacBook-Pro-2.local>