Make CloseAndGetPendingState async to allow blob work to complete (#16536)
https://dev.azure.com/fluidframework/internal/_workitems/edit/4554

## Description

closeAndGetPendingLocalState needs to wait until blobs are attached to
close. This is because there is no use on stashing blobs without knowing
user intention. The promise resolves if there are no non-attached blobs
in the pending list or when all pending blobs are attached via
"allBlobsAttached" event.

TODO
- [x] Complete test coverage (close while blobs are uploading and
waiting to round trip).
- [ ] Probably block further blob uploading while closing.
- [x] "cannot capture the pending local state during ordersequentially"
and similar tests are failing. Working on fixing all broken tests