Not register summarizer for visibility change (#16649)
## Description

There is currently a leak on teams ([Bug
8036636](https://office.visualstudio.com/OC/_workitems/edit/8036636)) in
which we register for visibility changes on the summarizer container and
we never unregister. We are investigating to understand whether this
leak could be avoided but the fact is that we do not need to register
for visibility changes when in the summarizer.