Deprecate some request related APIs (#16506)
# Description

There are many places in code that utilize the "request" pattern but
aren't used by external applications. This pattern should be deprecated
from these places to facilitate the full removal of the "request"
pattern.


[AB#4983](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/4983)

# Upcoming changes

### Request APIs deprecated from many places

The `request` API (associated with the `IFluidRouter` interface) has
been deprecated on a number of classes and interfaces. The following are
impacted:

-   `IRuntime` and `ContainerRuntime`
-   `IFluidDataStoreRuntime` and `FluidDataStoreRuntime`
-   `IFluidDataStoreChannel`
-   `MockFluidDataStoreRuntime`
-   `TestFluidObject`

Please migrate usage to the corresponding `entryPoint` or
`getEntryPoint()` of the object. The value for these "entryPoint"
related APIs is determined from factories (for `IRuntime` and
`IFluidDataStoreRuntime`) via the `initializeEntryPoint` method. If no
method is passed to the factory, the corresponding `entryPoint` and
`getEntryPoint()` will be undefined.

For an example implementation of `initializeEntryPoint`, see
[pureDataObjectFactory.ts](../packages/framework/aqueduct/src/data-object-factories/pureDataObjectFactory.ts#L83).

More information of the migration off the request pattern, and current
status of its removal, is documented in
[Removing-IFluidRouter.md](https://github.com/microsoft/FluidFramework/blob/main/packages/common/core-interfaces/Removing-IFluidRouter.md).