feat(bump:deps): Add experimental homegrown update checker (#16356)
The current implementation of the bump:deps command uses
npm-check-updates to do the checking of npm and to update the
package.json files. This was implemented for expediency, but the
implementation is slow and does far more work than it needs to.

To address these limitations, I have added an experimental homegrown
version updater. It's simpler overall, and much faster, but it needs
more testing. It's behind a new CLI flag (`--updateChecker homegrown`),
so the default behavior of the command won't change.

One new feature in this updater is that it will use exact dependencies
when setting dependencies on pre-release versions. With prerelease, we
don't want any semver range behavior - we want exact deps.