docs(tree): Reorganize and fix links (#16661)
## Description

Reorganize docs into more general folders based on the status of the doc
and rename them to snake case in order to make them consistent and
easier to link to.

## Reviewer Guidance

The review process is outlined on [this wiki
page](https://github.com/microsoft/FluidFramework/wiki/PR-Guidelines#guidelines).

The purpose of this reorganization is to rely less on folders for
organization which can lead to links breaking when we want to reorganize
docs in the future. This instead encourages organization using links and
landing pages. A temporary landing page is added as the README for the
docs folder to act as an example but doesn't actually establish any
organization yet. Organizing using links in pages allows context about
the relationship between links to be included and allows pages to be
included in multiple categories/landing pages rather than just a single
folder hierarchy. Changing these categories is also much easier and
doesn't break links.

Future improvements:
- Publish the docs as a Github Pages website
- Include backlinks (links to pages that link to the current page) to
improve connections between pages
- Split docs up into more atomic pages to increase linking and shared
documentation
- Publish API docs as markdown
- Clean up the readmes in the submodule folders to be minimal and link
to the corresponding page in the docs folder
- Add a linter to detect broken local links