Add getAliasDataStore API and deprecated getRootDataStore (#16558)
## Description
Added a `getAliasDataStore` API that returns an aliased data store with
the given alias. It throws a usage error if the data store doesn't exist
or is not aliased.

We currently have a `getRootDataStore` API that has been deprecated. It
returns an aliased data store but it is missing a few things:
- It returns an IFluidRouter. We are in the process of removing
IFluidRouter and all usages of request APIs.
- The API accepts a wait param which should be removed.
- We have removed the term and concept of "root" from the public data
store APIs. There is only aliasing now which translates internally to a
root data strore. So, the getter API should not have the term "root"
either.

## Reviewer guidance
I would love feedback on the following in particular - Is omitting the
`wait` flag fine. `getRootDataStore` has a wait param whose purpose is
to be able to wait for a "named" data store to show up if it doesn't
exist yet. This does not seem like a great use case and if needed it can
be implemented on top of this API by applications.


[AB#3459](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/3459)