AzureClient: Add JUnit reporting option to Azure End-to-end Tests (#16667)
## Description

Similar to other E2E tests in the FF repo, it would be nice to have
JUnit reports output from the Azure End-to-end test suite. With this
change, setting `FLUID_TEST_REPORT=1` causes tests to output a
nyc/junit-report.xml file.

Additionally, mocha-test-setup adds the ability for us to easily plug in
a test logger (e.g. ff-internal/aria-logger) for better log output
reporting. I added `getTestLogger` as the logger for AzureClient when
the mockLogger is not defined. As far as I can tell, there is currently
no usage of MockLogger in these tests, so this shouldn't change
anything. In the future, we can use a mix-in to output to both as
needed, I think.