[next] Deprecate IContainer.request and IDataStore.request, except for {url: "/" } (#16660)
## Description

We are working on deprecating `IFluidRouter` and `request` - see
[Removing-IFluidRouter.md](https://github.com/microsoft/FluidFramework/blob/main/packages/common/core-interfaces/Removing-IFluidRouter.md)
for more information of the migration off the request pattern, and
current status of its removal.

`IContainer` and `IDataStore` are "special" in that they are the APIs
that expose entryPoint. In some known consumers of FF, the
ContainerRuntime and DataStoreRuntimes are set up with a
`requestHandler` that routes requests for "/" to a the entryPoint object
(aka "root"). Since the entryPoint API is not fully rolled out /
mandatory yet, we can leave `request({ url: "/" })` as a temporary
alternative to access the entryPoint, while deprecating support for
requesting arbitrary URLs. After all, this is the problematic pattern we
are trying to discourage.

Planning to deprecate this accommodation in the next major release,
where entryPoint APIs will be mandatory and can be used directly.