tree2: Separate Struct Map and Leaf in ViewSchema (#16672)
## Description

As a step toward providing distinct APIs for Struct, Map and Leaf nodes
in the EditableTree and schema aware APIs, this separates some of the
ViewSchema types for these.

The main change here is that struct nodes no longer have a dummy
mapFields value that indicates empty fields, but instead use `undefined`
in a few places. This makes it possible to separate empty struct schema
from empty map schema, ensuring its always possible to select the
correct API to present (which is not yet implemented).

## Breaking Changes

Users of mapFields should expect undefined instead of emptyField for
non-map nodes.