Require leaf nodes to have a value (#16715)
## Description

As part of Schema2: adjust ValueSchema so that leaf nodes always have a
value (undefined is no longer allowed for ValueSchema.Serializeable).
This makes encoding into json simple (since json does not support
undefined). Any schema that requires this degree of freedom can make the
parent of the leaf node either store the leaf in a optional field, or
permit some other type there (such as an empty struct) to represent that
case. As far as I can tell none of our schema used this degree of
freedom, so none need updates.

## Breaking Changes

Some schema usage mad need to be adjusted for the above mentioned
changes.