Azure E2E: ContainerLoadStats event changed (#16724)
## Description

The "ContainerLoadStats" event now has additional context
"ContainerRuntime" in the eventName. This broke the Azure E2E test
suite.

Instead of just updating to the new eventName, I changed the search to
look for `endswith("ContainerLoadStats")` which should be more resilient
to future changes.

### Bonus

When `MockLogger` and `getTestLogger()` were defined, only `MockLogger`
would receive logs. This PR adds the ability to log to both using
`MultiSinkLogger` when necessary.