Azure scenario-runner cleanup (#16541)
## Description

Follow-up on #15057 addressing feedback from @ssimic2.

## Breaking Changes

Scenario-runner is experimental now, after #15057, so this should be OK.

1. All connection and service auth information comes from Env now,
instead of each individual scenario config.
2. All *RunnerClient files have been consolidated into
`scenarioRunnerClient.ts`, so running via client file would now be
`./dist/scenarioRunnerClient.js <ScenarioNameRunner> [...args]`

## Reviewer Guidance

My goal here is to reduce the amount of boilerplate required to create
more runners for various scenarios. Please let me know if this is a good
start or if I went overboard on deduping common configs.

Also, AzureClientRunner is kinda weird and doesn't fit the pattern of
the other Runners. It is likely **always** going to be the first stage
in any scenario config. Can we just make it a function that happens in
TestOrchestrator?