Ephemeral container - Server changes (#16624)
## Description

Ephemeral containers are containers whose artifacts (ops, summaries and
metadata) are stored only for the lifetime of the session. The container
lifecycle is completely controlled by the fluid server (for durable
containers, the deletes are client initiated). We make use of the
comparatively shorter lifetime of these containers to store their
artifacts closer to the user without breaking compliance which helps
these containers have better latency and reliability while using less
cogs.

## Reviewer Guidance

The flow is as follows:

- The isEphemeralContainer flag is sent by the driver as a part of the
create document flow #16006
- When this call reaches alfred, we store the flag as a part of the
document metadata (storage.ts) and pass it down to historian
- In historian, this flag is cached in redis, and is passed further down
to Gitrest. For the subsequent calls to historian in this session, the
info about whether the container is ephemeral is fetched from the cache
and is passed onto Gitrest
- In Gitrest, when using isomorphic git, we pass in two
IFileSystemManagerFactories. Depending on the isEphemeralContainer flag,
we choose the appropriate IFileSystemManagerFactory

By default, the IFileSystemManagerFactory for ephemeral containers will
be RedisFsManagerFactory. Redis was introduced as a backing filesystem
for isomorphic git as a part of this PR #16470

Some more points to note:
The ephemeral flag is passed from one microservice to another (Alfred ->
Historian -> Gitrest) only if the flag is set to true. It is not passed
in the cases where it is not an ephemeral container (if driver set this
flag to false or left it undefined), to reduce to size of the request
header and also to make it easier to control turning this feature on and
off