Improve trunk eviction in EditManager (#16713)
## Description

The SharedTree EditManager's trunk commit eviction was both more
complicated than necessary and also buggy. This PR cleans up the code
and addresses the known bugs. Previously, when the minimum sequence
number advanced to some sequence number _M_, all commits with sequence
numbers below _M_ were dropped, but a commit with exactly sequence
number _M_ was not because any number of branches might be based off of
that commit (e.g. the local branch), and dropping it was impossible
because a branch might continue to reference it. After this PR, that
commit is still not dropped from memory, however it takes the place of
the special "trunk base" commit or "origin" commit which precedes the
trunk but is not considered part of the trunk. Not only does this remove
multiple edge cases from the eviction logic (and the bugs that currently
dwell therein), it also makes it more intuitive to predict which commits
are removed from the trunk as the minimum sequence number advances.


![image](https://github.com/microsoft/FluidFramework/assets/78610362/0d0e3c3b-9d93-4bed-bba6-b239f618f95f)

This PR also does some minor cleanup and optimizations in EditManager,
adds some more eviction test cases, and enables an eviction pass after
tracked branches rebase (which will evict the trunk more aggressively in
scenarios where a long-lived user-created branch repeatedly rebases over
a parent branch).