added approveProposalComplete event to QuorumProposal (#16765)
https://dev.azure.com/fluidframework/internal/_workitems/edit/4749/
The goal of this item is to remove the usage of the deltaManager in the
sameContainerMigrationTool. The root cause of the migration tool's
reliance on the deltaManager is the lack of conflict resolution and
eventing in the QuorumProposal.
Improving the conflict resolution and eventing in the QuorumProposal
will provide the migration tool with information regarding the proposal
state without having to spy on the op stream.

This PR covers the first segment addressing the conflict resolution and
eventing. It checks that a pending proposal is unique before emitting an
"approveProposalComplete" event. By verifying that the proposal is
unique, we can be certain that there won't be conflicts after approval.

A follow up PR will consume the updated protocol-base and update the
sameContainerMigrationTool to make use of this new eventing.