Simplify summarizable implementations (#16758)
## Description

This removes the reusable blob pattern for `ForestSummarizer` and
`EditManagerSummarizer`, which are unlikely to ever have the exact same
summary data and thus unlikely to benefit from it. They also didn't
correctly implement `load` anyway. Conveniently, this cleans up bad
JSON.stringify usages which don't behave well with GC.

Removes double-stringification from `ForestSummarizer`.

I converted all summarizables to use `createSingleBlobSummary` when
possible.

Lastly, this modifies SharedTreeCore's `getGCData` implementation to
delegate to the base implementation and extend it with the
summarizables' implementations, which should make it correct for the
time being (though we still have a testing gap here).


[AB#5062](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/5062)