fix(typetests): Don't fail when packages have no dependencies (#16717)
Type test preparation was failing when trying to update the
devDependencies of a package without the devDependencies property in
package.json. I fixed this by doing an undefined check.

I also tightened the typing of our PackageJson type by making
dependencies and devDependencies nullable properties. This did require a
few updates to handle the undefined checks, but overall that was a small
price to pay for better type safety.

Finally, I added a catch handler to the code that executes the
per-package function. This gives us an opportunity to log the package
that's causing the error along with the stack trace. This will make
future debugging easier.