tree2: Defer schema compatibility error, and more testing (#16782)
## Description

Add some collab tests and test for errors for schematize.
Defer errors from schema changing to an incompatible one until
afterBatch. This is enough to make the tests pass.

This addresses an issue found when refactoring some tests to use
schematize to initialize the tree, which, if using a value field for the
root, failed when running processMessages, even when not collaborating.