stop checkpoints when document is corrupt (#16802)
Currently, some idleTime checkpoints have been occurring after the
document is marked as corrupt on in the document partition due to errors
being thrown from the scribe lambda. This leads to assertion errors as
Kafka checkpoints when marking a document as corrupt, and the assertion
that the offset to be checkpointed is between the last checkpointed
offset and the largest offset processed by Kafka fails.

To avoid this, we need to stop idleTime checkpointing when we throw
errors that lead to documents being marked as corrupt.