Removed unused summarization attempts code and moved validate summary (#16804)
The summarization retry logic was updated to do only 2 attempts by
#14164. It added a feature flag to change the attempt count if needed.
The feature flag was never used and the 2 attempts are the default now.
Removed the code that deals with the extra summarization attempts.

Also, moved the summary validation in container runtime's
`submitSummary` function to right after generating the summary. The code
in between the summary generation and validation does not need to run.
The properties from the generated summary don't need to be part of the
failed validation result because the summary is incorrect so any state
from it shouldn't be exposed.